implementação do front

+++ ImersaoDevBack/Models/Aluno.cs
@@ -5,10 +5,10 @@ namespace ImersaoDev.models
     {
         public string? Name { get; set; }
         public string? Materia { get; set; }
-        public double NotaUnidade1 { get; set; }
-        public double NotaUnidade2 { get; set; }
-        public double NotaUnidade3 { get; set; }
-        public double media { get; set; }
+        public float NotaUnidade1 { get; set; }
+        public float NotaUnidade2 { get; set; }
+        public float NotaUnidade3 { get; set; }
+        public float media { get; set; }
         
         public char? conceito { get; set; }
     }
